+++ src/Layers/xrRender/r__dsgraph_build.cpp
@@ -701,7 +701,7 @@ void R_dsgraph_structure::build_subspace()
         Fvector box_radius;
         box_radius.set(o.query_box_side, o.query_box_side, o.query_box_side);
         Sectors_xrc.box_query(CDB::OPT_FULL_TEST, RImplementation.rmPortals, o.view_pos, box_radius);
-        for (int K = 0; K < Sectors_xrc.r_count(); K++)
+        for (size_t K = 0; K < Sectors_xrc.r_count(); K++)
         {
             CPortal* pPortal = Portals[RImplementation.rmPortals->get_tris()[Sectors_xrc.r_begin()[K].id].dummy];
             pPortal->bDualRender = TRUE;
