+++ src/Layers/xrRenderPC_R1/FStaticRender.cpp
@@ -485,7 +485,7 @@ void CRender::Calculate()
             Fvector box_radius;
             box_radius.set(EPS_L * 2, EPS_L * 2, EPS_L * 2);
             dsgraph.Sectors_xrc.box_query(CDB::OPT_FULL_TEST, rmPortals, Device.vCameraPosition, box_radius);
-            for (int K = 0; K < dsgraph.Sectors_xrc.r_count(); K++)
+            for (size_t K = 0; K < dsgraph.Sectors_xrc.r_count(); K++)
             {
                 CPortal* pPortal = dsgraph.Portals[rmPortals->get_tris()[dsgraph.Sectors_xrc.r_begin()[K].id].dummy];
                 pPortal->bDualRender = TRUE;
